chore: deleted icon.tsx
