--- apps/web/src/components/Assets/Icons.tsx
@@ -1,81 +1,0 @@
-type IconProps = React.HTMLAttributes<SVGElement>;
-
-export const Icons = {
-  DarkLogo: (props: IconProps) => (
-    <svg
-      width="33"
-      height="32"
-      viewBox="0 0 33 32"
-      {...props}
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <rect x="0.5" width="32" height="32" rx="16" fill="#23031B" />
-      <path
-        d="M6.43523 16.0764C6.07601 16.3662 5.86719 16.803 5.86719 17.2645C5.86719 18.5474 7.35366 19.2581 8.35213 18.4526L9.45598 17.5621L7.42998 15.2739L6.43523 16.0764Z"
-        fill="#FF7F55"
-      />
-      <path
-        d="M8.42473 14.4715L7.42998 15.2739L9.45598 17.5621L10.5598 16.6716C10.8278 16.4554 11.1267 16.293 11.4414 16.186L9.45598 13.7879C9.0961 13.9806 8.75063 14.2085 8.42473 14.4715Z"
-        fill="#FFEDB9"
-      />
-      <path
-        d="M11.4414 16.186C12.1626 15.9408 12.9666 15.9867 13.6708 16.3428L14.3833 13.2594C12.7418 12.7917 10.9712 12.9768 9.45598 13.7879L11.4414 16.186Z"
-        fill="#FF7F55"
-      />
-      <path
-        d="M13.6708 16.3428L16.4994 17.7734L17.9138 14.979L16.4994 14.2637L15.9622 13.9426C15.4595 13.6422 14.9286 13.4148 14.3833 13.2594L13.6708 16.3428Z"
-        fill="#FFEDB9"
-      />
-      <path
-        d="M16.4994 17.7734L17.0367 18.0945C17.6811 18.4797 18.3722 18.7448 19.0801 18.892L20.1273 15.9607C19.8535 15.9123 19.5842 15.8238 19.3281 15.6943L17.9138 14.979L16.4994 17.7734Z"
-        fill="#FFEDB9"
-      />
-      <path
-        d="M20.1273 15.9607L19.0801 18.892C20.4373 19.1742 21.8566 19.0228 23.1286 18.4526L21.8672 15.7253C21.325 15.9831 20.716 16.0647 20.1273 15.9607Z"
-        fill="#CEB66F"
-      />
-      <path
-        d="M22.4391 15.3655C22.2607 15.5094 22.0685 15.6295 21.8672 15.7253L23.1286 18.4526C23.6387 18.224 24.125 17.928 24.5742 17.5657L25.5689 16.7632L23.5429 14.475L22.4391 15.3655Z"
-        fill="#FFEDB9"
-      />
-      <path
-        d="M26.5637 15.9607C26.9229 15.6709 27.1317 15.2341 27.1317 14.7726C27.1317 13.4897 25.6452 12.779 24.6468 13.5845L23.5429 14.475L25.5689 16.7632L26.5637 15.9607Z"
-        fill="#CFB76E"
-      />
-    </svg>
-  ),
-  Database: (props: IconProps) => (
-    <svg
-      width="18"
-      height="18"
-      viewBox="0 0 18 18"
-      fill="none"
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M9 9C12.4173 9 15.1875 7.48896 15.1875 5.625C15.1875 3.76104 12.4173 2.25 9 2.25C5.58274 2.25 2.8125 3.76104 2.8125 5.625C2.8125 7.48896 5.58274 9 9 9Z"
-        fill="#A6554D"
-        stroke="#A6554D"
-        strokeWidth="1.125"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M2.8125 5.625V9C2.8125 10.864 5.58281 12.375 9 12.375C12.4172 12.375 15.1875 10.864 15.1875 9V5.625"
-        stroke="#A6554D"
-        strokeWidth="1.125"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M2.8125 9V12.375C2.8125 14.239 5.58281 15.75 9 15.75C12.4172 15.75 15.1875 14.239 15.1875 12.375V9"
-        stroke="#A6554D"
-        strokeWidth="1.125"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  ),
-};
