+++ apps/web/src/components/AuthUI/SignIn.tsx
@@ -5,14 +5,14 @@ import { useSearchParams } from "next/navigation";
 
 
 import { SocialLogin } from "./SocialLogin";
-import { Icons } from "../Assets/Icons";
+import { SandwormDarkLogo } from "../Assets/SandwormDarkLogo";
 
 export const SignIn = () => {
   const params = useSearchParams();
   const error = params?.get("error");
   return (
     <div className="w-full max-w-md rounded-lg border border-[#ffffff30] p-6 content flex flex-col justify-center items-center h-full mx-auto">
-      <Icons.DarkLogo />
+      <SandwormDarkLogo />
 
       <h2 className="text-3xl font-medium text-center text-black mb-6 mt-4 roobert">
         Welcome back to <span className="uppercase font-bold">SANDWORM</span>!
